@@ -1,19 +1,19 @@
 package cmd
 
 import (
-	"errors"
+	"context"
 
+	"github.com/cockroachdb/errors"
 	"github.com/spf13/cobra"
 
-	"narou/adapter/logger"
-	"narou/adapter/repository/epub"
-	metadataRepo "narou/adapter/repository/metadata"
-	"narou/adapter/repository/novel"
 	"narou/config"
+	"narou/domain/epub"
+	"narou/domain/metadata"
+	"narou/domain/novel"
 	"narou/infrastructure/database"
 	"narou/infrastructure/storage"
-	"narou/interface/controller"
-	"narou/usecase/interactor/convert"
+	"narou/sdk/logger"
+	"narou/usecase/convert"
 )
 
 // convertCmd represents the convert command
@@ -22,25 +22,28 @@ var convertCmd = &cobra.Command{
 	Short: "convert html to epub",
 	RunE: func(c *cobra.Command, args []string) error {
 		if len(args) == 0 {
-			return errors.New("need id")
+			return errors.Wrapf(ErrRequiredArgsNotFound, "command:%s", "convert")
 		}
 		ctx := c.Context()
-		lg := logger.NewLogger(ctx)
+		ctx = context.WithValue(ctx, "batch-name", "convert")
+		lg, err := logger.NewCLILogger(ctx)
+		if err != nil {
+			return err
+		}
 		db, err := database.GetConn()
 
 		if err != nil {
-			panic(err)
+			return errors.Wrapf(err, "db connection fail. command:%s", "convert")
 		}
 		cfg := config.GetConfigure()
-		cvt := convert.NewConvertInteractor(ctx,
+		useCase := convert.NewConvertUseCase(
 			lg,
-			metadataRepo.NewRepository(db),
+			metadata.NewRepository(db),
 			novel.NewRepository(storage.NewManager(cfg)),
 			epub.NewRepository(storage.NewManager(cfg)),
 			cfg,
 			DependencyInjection(args[0]))
-
-		return controller.NewConvertController(cvt, lg).Execute(args)
+		return useCase.Execute(ctx, metadata.URI(args[0]))
 	},
 }
 
