@@ -5,10 +5,11 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/cockroachdb/errors"
 	"github.com/labstack/echo/v4"
 
-	"narou/adapter/logger"
 	"narou/infrastructure/waf/grpc_client/middleware"
+	"narou/sdk/logger"
 	pb "narou/usecase/port/boudary/proto/novel"
 )
 
@@ -16,7 +17,10 @@ func GetList(c echo.Context) error {
 	ctx, canceler := context.WithTimeout(c.Request().Context(), time.Second*1)
 	defer canceler()
 
-	lg := logger.NewLogger(ctx)
+	lg, err := logger.NewServerLogger(ctx)
+	if err != nil {
+		return errors.Wrapf(err, "logger err")
+	}
 	lg.Info("grpc grpc_client get start")
 	lg.Info("ctx", "ctx", ctx)
 	defer lg.Info("grpc grpc_client get end")
