@@ -2,15 +2,15 @@ package download
 
 import "narou/domain/metadata"
 
-type download struct {
+type Download struct {
 }
 
-func NewOutPutBoundary() *download {
-	return &download{}
+func NewOutPutBoundary() *Download {
+	return &Download{}
 }
 
-//  OutPUtBoundary ダウンロード後に表示するデータ
-func (download) OutPUtBoundary(novels []metadata.Novel) []string {
+// OutPutBoundary ダウンロード後に表示するデータ
+func (Download) OutPutBoundary(novels []metadata.Novel) []string {
 	var ret []string
 	for _, v := range novels {
 		ret = append(ret, v.Title)
